Position and orientation units changed to mm and rad. Updated indentations.

+++ src/antrobot_drive.h
@@ -6,6 +6,6 @@
 
 #include "rs232.h"
 
-int send_command(std::string command, std::vector<int> *position);
+int send_command(std::string command, std::vector<double> *position);
 int init_comport(int port_number);
 int close_comport(int port_number);
